types: restructure + move core / data source related types
