+++ types/src/front/assistant/conversation.ts
@@ -1,9 +1,9 @@
-import { UserType, WorkspaceType } from "../../../front/types/user";
-import { ModelId } from "../../../shared/model_id";
+import { UserType, WorkspaceType } from "../../front/user";
+import { ModelId } from "../../shared/model_id";
 
-import { DustAppRunActionType } from "../../../front/types/assistant/actions/dust_app_run";
-import { RetrievalActionType } from "../../../front/types/assistant/actions/retrieval";
-import { AgentConfigurationType } from "../../../front/types/assistant/agent";
+import { DustAppRunActionType } from "../../front/assistant/actions/dust_app_run";
+import { RetrievalActionType } from "../../front/assistant/actions/retrieval";
+import { AgentConfigurationType } from "../../front/assistant/agent";
 
 /**
  * Mentions
