+++ types/src/front/data_source.ts
@@ -1,4 +1,4 @@
-import { ModelId } from "../../shared/model_id";
+import { ModelId } from "../shared/model_id";
 
 export type DataSourceVisibility = "public" | "private";
