+++ types/src/index.ts
@@ -1,12 +1,13 @@
-export * from "./front/types/key";
-
-export * from "./front/api_handlers/public/assistant/types";
-export * from "./front/types/data_source";
-
-export * from "./front/types/user";
 export * from "./shared/model_id";
-export * from "./front/types/assistant/agent";
-export * from "./front/types/assistant/actions/retrieval";
-export * from "./front/types/assistant/conversation";
+
+export * from "./core/data_source";
+
 export * from "./front/lib/supported_models";
-export * from "./front/types/assistant/actions/dust_app_run";
+export * from "./front/key";
+export * from "./front/data_source";
+export * from "./front/user";
+export * from "./front/assistant/agent";
+export * from "./front/assistant/actions/retrieval";
+export * from "./front/assistant/conversation";
+export * from "./front/assistant/actions/dust_app_run";
+export * from "./front/api_handlers/public/assistant";
